travis-specific-requirements.txt
reprint
tabulate>=0.8.2,<1.0
cfn_lint>=0.13.0,<1.0
setuptools>=40.4.3
boto3>=1.9.21,<2.0
botocore>=1.12.21,<2.0
yattag>=1.10.0,<2.0
PyYAML~=5.1
jinja2>=2.10.0,<3.0
requests>2.17.0
jsonschema~=3.0
docker==4.2.2
dulwich~=0.19
dataclasses;python_version<"3.7"
dataclasses-jsonschema>=2.9.0,<3.0
pip
